+++ packages/client/src/components/modelBrowser/index.tsx
@@ -239,7 +239,7 @@ export const ModelBrowser: React.FC<Props> = observer(
             <IconButton
               title="Refresh"
               onClick={() => {
-                modelStore.refresh();
+                modelStore.refresh(true);
               }}
             >
               <BsArrowClockwise />